myems-api/requirements.txt
anytree
simplejson
mysql-connector-python
falcon
falcon_cors
falcon-multipart
gunicorn
et_xmlfile
jdcal
openpyxl
pillow
python-decouple